pyproject.toml
[project]
name = "jepa"
version = "0.1.0"
description = "JEPA-based clip mining for autonomous driving"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Research Team" }
]

dependencies = [
    # Core Deep Learning
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    
    # Scientific Computing
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    
    # Image Processing
    "pillow>=10.0.0",
    
    # HuggingFace Ecosystem
    "huggingface-hub>=0.20.0",
    "tokenizers>=0.15.0",
    "safetensors>=0.4.0",
    
    # Utilities
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    
    # File System
    "fsspec>=2023.6.0",
    "filelock>=3.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "ipykernel>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jepa"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "ipykernel>=6.0.0",
]
